Don't delete target directory when it doesn't exist.

+++ src/zeit/releaser/aftercheckout.py
@@ -14,8 +14,9 @@ def copy_unstaged_sources(data):
 
 
 def copy_js_css(src, dest, target):
-    print 'deleting %s/%s' % (dest, target)
-    shutil.rmtree('%s/%s' % (dest, target))
+    if os.path.exists(target):
+        print 'deleting %s/%s' % (dest, target)
+        shutil.rmtree('%s/%s' % (dest, target))
     print 'copying %s/%s to %s/%s' % (src, target, dest, target)
     shutil.copytree('%s/%s' % (src, target), '%s/%s' % (dest, target))
